+++ frontend/providers/dbprovider/src/components/QuotaBox/index.tsx
@@ -51,18 +51,20 @@ ${t('common.Surplus')}: ${(limit - used).toFixed(2)} ${unit}`;
     <Box borderRadius={'md'} border={showBorder && theme.borders.base} bg={'#FFF'}>
       <Box
         py={3}
-        px={4}
+        px={'20px'}
         borderBottom={showBorder && theme.borders.base}
         color={'grayModern.900'}
         fontWeight={500}
       >
         {t('app.Resource Quota')}
       </Box>
-      <Flex flexDirection={'column'} gap={'14px'} py={3} px={4}>
+      <Flex flexDirection={'column'} gap={'14px'} py={'16px'} px={'20px'}>
         {quotaList.map((item) => (
           <MyTooltip key={item.type} label={item.tip} placement={'top-end'} lineHeight={1.7}>
             <Flex alignItems={'center'}>
-              <Box flex={'0 0 60px'}>{t(item.type)}</Box>
+              <Box flex={'0 0 60px'} textTransform={'capitalize'}>
+                {t(item.type === 'cpu' ? 'CPU' : item.type)}
+              </Box>
               <Progress
                 flex={'1 0 0'}
                 borderRadius={'sm'}
