+++ frontend/providers/dbprovider/src/pages/db/detail/components/DumpImport/index.tsx
@@ -198,30 +198,9 @@ export default function DumpImport({ db }: { db?: DBDetailType }) {
     <Box h={'100%'} position={'relative'}>
       <Flex borderTop={'1px solid #EAEBF0'} h="100%">
         <Box flex={'0 1 256px'} borderRight={'1px solid #EAEBF0'}>
-          <QuotaBox showBorder={false} />
-          {/* {db && (
-            <PriceBox
-              components={[
-                {
-                  cpu: db?.cpu,
-                  memory: db?.memory,
-                  storage: db?.storage,
-                  replicas: [db?.replicas || 1, db?.replicas || 1]
-                },
-                ...(db?.dbType === DBTypeEnum.redis
-                  ? (() => {
-                      const config = RedisHAConfig(db?.replicas > 1);
-                      return [
-                        {
-                          ...config,
-                          replicas: [config.replicas, config.replicas]
-                        }
-                      ];
-                    })()
-                  : [])
-              ]}
-            />
-          )} */}
+          <Box px={'4px'} pt={'14px'}>
+            <QuotaBox showBorder={false} />
+          </Box>
         </Box>
         <Box flex={1} pt="35px" px="40px" overflowY={'auto'}>
           <Text fontSize={'base'} fontWeight={'bold'} color={'grayModern.900'}>
